apps/api-gateway/requirements.txt
Django==5.2.12
djangorestframework==3.16.1
strawberry-graphql[django]==0.284.1
httpx==0.28.1
django-cors-headers==4.9.0
PyJWT==2.12.1
cryptography==46.0.5
python-dotenv==1.1.1
gunicorn==23.0.0
pytest==8.4.2
pytest-django==4.11.1